reconciliation function is refactored a bit
