+++ lib/dre/index.js
@@ -5,9 +5,19 @@ var bbMatch = require("blue-button-match");
 var _ = require("underscore");
 var async = require("async");
 
+//refactored reconciliation functions
+var demographics = require('./demographics');
+var dedup_new = require('./dedup_new');
+var remove_dups = require('./remove_dups');
+
 //Main function, performs match and dedupes.
 
 function reconcile(username, newObject, baseObject, newRecordID, callback) {
+    //HACKETY HACK, skipping reconciliation for now
+    callback(null, null, null);
+
+
+
     newObjectForParsing = newObject;
 
     var baseObjectForParsing = {};
@@ -20,26 +30,7 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
         }
     }
 
-    //BB Matching library expects object for demographics.
-    function prepDemographics() {
-        if (baseObjectForParsing.demographics instanceof Array) {
-            if (baseObjectForParsing.demographics.length > 0) {
-                baseObjectForParsing.demographics = baseObjectForParsing.demographics[0];
-
-                //remove PIM artifacts in demographics section
-                delete baseObjectForParsing.demographics.pim;
-            }
-        }
-        if (newObjectForParsing.demographics instanceof Array) {
-            if (newObjectForParsing.demographics.length > 0) {
-                newObjectForParsing.demographics = newObjectForParsing.demographics[0];
-
-                //remove PIM artifacts in demographics section
-                delete newObjectForParsing.demographics.pim;
-            }
-        }
-    }
-    prepDemographics();
+    demographics.prepDemographics(baseObjectForParsing, newObjectForParsing);
 
     baseObjectForParsing = {}.data = baseObjectForParsing;
     newObjectForParsing = {}.data = newObjectForParsing;
@@ -53,66 +44,18 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
     delete baseObjectForParsing.data;
     delete newObjectForParsing.data;
 
-    function revertDemographics() {
-        if (_.isObject(newObjectForParsing.demographics) === true && _.isArray(newObjectForParsing.demographics) === false) {
-            newObjectForParsing.demographics = new Array(newObjectForParsing.demographics);
-        }
-    }
 
-    function deDuplicateNew(match, matchObject) {
-        _.map(match, function (value, key) {
-            if (_.isArray(value)) {
-                //Find all duplicate source entries.
-                var duplicateArray = [];
-                for (var deLoop in value) {
-                    if (value[deLoop].dest === 'src' && value[deLoop].match === 'duplicate') {
-                        duplicateArray.push(value[deLoop]);
-                    };
-                }
 
-                //Find intersection and drop one.
-                for (var srcLoop in duplicateArray) {
-                    for (var destLoop in duplicateArray) {
 
-                        if (duplicateArray[srcLoop].src_id === duplicateArray[destLoop].dest_id) {
-                            if (duplicateArray[srcLoop].dest_id === duplicateArray[destLoop].src_id) {
-                                duplicateArray.splice(destLoop, 1);
-                                matchObject[key].splice(destLoop, 1);
-                            }
-                        }
-                    }
-                }
-
-                //Remove remaining entries from source.
-                for (var iSrc in value) {
-                    for (var iDest in duplicateArray) {
-                        if (_.isEqual(value[iSrc], duplicateArray[iDest])) {
-                            value.splice(iSrc, 1);
-                        }
-                    }
-                }
-
-            }
-
-        });
-
-        var returnObject = {
-            match: match,
-            new_entries: matchObject
-
-        }
-
-        return returnObject;
-    }
 
     function removeSourceMatches(match) {
 
         var returnMatch = {};
 
-        _.map(match, function (value, matchKey) {
+        _.map(match, function(value, matchKey) {
 
             if (matchKey !== 'demographics') {
-                var nonSourceMatch = _.filter(match[matchKey], function (object, objectIndex) {
+                var nonSourceMatch = _.filter(match[matchKey], function(object, objectIndex) {
                     if (object.dest === 'src') {
                         return false;
                     } else {
@@ -135,144 +78,7 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
 
     }
 
-    //Splice duplicate entries from input array.
-    //Add source attribution to matched entries.
-    function removeDuplicates(username, match, newRecord, baseRecord, recordID) {
 
-        var newRecord = _.clone(newRecord);
-
-        var duplicateArray = [];
-
-        var newMatch = {};
-
-        _.map(match, function (value, matchKey) {
-            if (_.isArray(value)) {
-                duplicateArray = _.where(value, {
-                    dest: 'dest',
-                    match: 'duplicate'
-                });
-
-                duplicateNewIndexArray = _.uniq(_.pluck(duplicateArray, 'src_id'));
-                duplicateBaseIndexArray = _.uniq(_.pluck(duplicateArray, 'dest_id'));
-
-                if (matchKey === 'social_history') {
-                    //console.log(match);
-                    //console.log(newRecord.social_history[0]);
-                    //console.log(baseRecord.social_history[4]);
-
-                }
-
-                var attributionRecord = _.filter(baseRecord[matchKey], function (object, objectIndex) {
-                    if (_.contains(duplicateBaseIndexArray, objectIndex.toString())) {
-                        return true;
-                    } else {
-                        return false;
-                    }
-                });
-
-                //Update Source attribution.
-                async.each(attributionRecord, function (attribution, callback) {
-                    record.duplicateEntry(matchKey, username, attribution._id, recordID, function (err) {
-                        if (err) {
-                            callback(err);
-                        } else {
-                            callback(null);
-                        }
-                    })
-                }, function (err) {
-                    if (err) {
-                        callback(err);
-                    } else {
-                        //callback();
-                    }
-                });
-
-                //Remove Duplicates from Entry List.
-                var returnRecord = _.filter(newRecord[matchKey], function (object, objectIndex) {
-                    if (_.contains(duplicateNewIndexArray, objectIndex.toString())) {
-                        return false;
-                    } else {
-                        return true;
-                    }
-                });
-
-                //Remove Duplicates from Match List.
-                //use duplicate new index array to go through matches, and remove.
-                var returnMatch = _.filter(match[matchKey], function (object, objectInex) {
-                    if (_.contains(duplicateNewIndexArray, object.src_id)) {
-                        return false;
-                    } else {
-                        return true;
-                    }
-                });
-
-                //console.log(returnMatch);
-
-                //console.log(returnRecord);
-
-                newRecord[matchKey] = returnRecord;
-
-                if (returnMatch === undefined) {
-                    returnMatch = [];
-                }
-
-                if (returnRecord === undefined) {
-                    returnRecord = [];
-                }
-
-                newMatch[matchKey] = returnMatch;
-
-            } else {
-                if (value.match === 'duplicate' && matchKey === 'demographics') {
-
-                    var attributionRecord = baseRecord[matchKey];
-
-                    async.each(attributionRecord, function (attribution, callback) {
-                        record.duplicateEntry(matchKey, username, attribution._id, recordID, function (err) {
-                            if (err) {
-                                callback(err);
-                            } else {
-                                callback(null);
-                            }
-                        })
-                    }, function (err) {
-                        if (err) {
-                            callback(err);
-                        } else {
-                            //callback();
-                        }
-                    });
-
-                    delete newRecord[matchKey];
-                    //delete newMatch[matchKey];
-                } else if (matchKey === 'demographics') {
-                    newMatch[matchKey] = value;
-                }
-
-            }
-        });
-
-        _.map(newRecord, function (value, field) {
-            if (_.isArray(value)) {
-                if (value.length === 0) {
-                    delete newRecord[field];
-                }
-            }
-        });
-
-        //console.log(match);
-
-        //console.log(JSON.stringify(newMatch, null, 10));
-
-        //This area can equal undefined on new_record entries for some reason.
-
-        var returnObject = {
-            new_record: newRecord,
-            new_match: newMatch
-        }
-
-        return returnObject;
-    }
 
     //duplicate matches must be purges from queue as well as entries!!!!
 
@@ -283,7 +89,7 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
 
         //console.log(match);
 
-        _.map(match, function (value, matchKey) {
+        _.map(match, function(value, matchKey) {
 
             outputPartialObjectArray[matchKey] = [];
             outputNewObjectArray[matchKey] = [];
@@ -303,9 +109,9 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
                     //console.log(baseObjectArray)
 
                     var partialOutput = {
-                                partial_entry: newObjectArray[matchKey][0],
-                                partial_match: match[matchKey],
-                                match_entry_id: baseObjectArray[matchKey][0]._id
+                        partial_entry: newObjectArray[matchKey][0],
+                        partial_match: match[matchKey],
+                        match_entry_id: baseObjectArray[matchKey][0]._id
                     };
 
                     var multiPartialOutput = {
@@ -339,10 +145,10 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
                 //if (matchKey === 'social_history') {
                 //    console.log(groupCheckedArray);    
                 //}
-                
 
-                _.each(groupCheckedArray, function (element, index) {
-                    var newElementArray = _.filter(element, function (elementItem, elementItemIndex) {
+
+                _.each(groupCheckedArray, function(element, index) {
+                    var newElementArray = _.filter(element, function(elementItem, elementItemIndex) {
                         if (elementItem.match === 'new') {
                             return true;
                         } else {
@@ -390,7 +196,7 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
 
                             //console.log(partialElementArray.length);
 
-                            _.each(partialElementArray, function (partialElement, partialIndex) {
+                            _.each(partialElementArray, function(partialElement, partialIndex) {
 
                                 if (matchKey === 'social_history') {
                                     //console.log(partialElementArray);
@@ -453,10 +259,10 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
 
     }
 
-    revertDemographics();
+    demographics.revertDemographics(newObjectForParsing);
 
     //Remove Overlapping Source Matches.
-    var deDuplicatedSourceRecords = deDuplicateNew(matchResult.match, newObject);
+    var deDuplicatedSourceRecords = dedup_new.deDuplicateNew(matchResult.match, newObject);
 
     //Remove All 'src' matches.  Currently not required.
     var nonSourceMatches = removeSourceMatches(deDuplicatedSourceRecords.match);
@@ -464,7 +270,7 @@ function reconcile(username, newObject, baseObject, newRecordID, callback) {
 
     //Remove Duplicates from save, update Record Entry.
 
-    var deDuplicatedNewRecord = removeDuplicates(username, nonSourceMatches.match, deDuplicatedSourceRecords.new_entries, baseObject, newRecordID);
+    var deDuplicatedNewRecord = remove_dups.removeDuplicates(username, nonSourceMatches.match, deDuplicatedSourceRecords.new_entries, baseObject, newRecordID);
     //console.log(JSON.stringify(deDuplicatedNewRecord.new_match, null, 10));
 
     //Split incoming entries into new/partial.
